*Fix a crash occuring after a product is selected and add the ability to change an existing product

+++ models/product.js
@@ -13,7 +13,8 @@ const getProductsfromfile = (cb) => {
 }
 
 module.exports = class Product {
-    constructor(title, url, price, description) {
+    constructor(id, title, url, price, description) {
+        this.id = id;
         this.title = title;
         this.imageUrl = url;
         this.price = price;
@@ -21,12 +22,25 @@ module.exports = class Product {
     }
 
     save(){ //save to products.json
-        this.id = Math.random().toString();
         getProductsfromfile(products => {
-            products.push(this);
-            fs.writeFile(filePath, JSON.stringify(products), (error) => {
-                console.log(error);
-            });
+            if(this.id) //if a product with the specified id exists
+            {
+                const existingProductIndex = products.findIndex(product => product.id === this.id);
+                const updatedProducts = [...products]; // pull out existing products and then store them in a new array
+                updatedProducts[existingProductIndex] = this;
+                fs.writeFile(filePath, JSON.stringify(updatedProducts), (error) => {
+                    console.log(error);
+                });
+            } 
+            else 
+            {
+                this.id = Math.random().toString();
+                products.push(this);
+                fs.writeFile(filePath, JSON.stringify(products), (error) => {
+                    console.log(error);
+                });
+            }
+
         });
         // fs.readFile(filePath, (error, fileContent) => {
         //     let products = [];
